requirements.txt
Django==2.0.2
py2neo==3.1.2
djangorestframework==3.7.7
